pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundation-model-any-ft"
version = "0.1.0"
description = "Fine-tune lightweight heads on frozen vision foundation models"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Foundation Model Any FT Contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "albumentations>=1.3.0",
    "Pillow>=9.0.0",
    "numpy>=1.24.0",
    "pycocotools>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
backend = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlmodel>=0.0.8",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
]
all = ["foundation-model-any-ft[backend,dev]"]

[project.scripts]
fft = "core.cli:main"

[tool.setuptools.packages.find]
include = ["core*", "sdk*", "backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py39"
